stock_market_analysis/requirements.txt
google-generativeai>=0.3.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1 
streamlit==1.17.0
textblob==0.17.1    
streamlit==1.17.0
yfinance==0.1.70
matplotlib==3.7.1
textblob==0.17.1
plotly==5.10.0
altair==4.2.0